Format /app.

+++ assets/javascripts/app/db.js
@@ -7,7 +7,7 @@
  * DS207: Consider shorter variations of null checks
  * Full docs: https://github.com/decaffeinate/decaffeinate/blob/master/docs/suggestions.md
  */
-(function() {
+(function () {
   let NAME = undefined;
   let VERSION = undefined;
   const Cls = (app.DB = class DB {
@@ -27,9 +27,15 @@
     }
 
     db(fn) {
-      if (!this.useIndexedDB) { return fn(); }
-      if (fn) { this.callbacks.push(fn); }
-      if (this.open) { return; }
+      if (!this.useIndexedDB) {
+        return fn();
+      }
+      if (fn) {
+        this.callbacks.push(fn);
+      }
+      if (this.open) {
+        return;
+      }
 
       try {
         this.open = true;
@@ -47,11 +53,15 @@
       const db = event.target.result;
 
       if (db.objectStoreNames.length === 0) {
-        try { db.close(); } catch (error1) {}
+        try {
+          db.close();
+        } catch (error1) {}
         this.open = false;
         this.fail('empty');
       } else if (error = this.buggyIDB(db)) {
-        try { db.close(); } catch (error2) {}
+        try {
+          db.close();
+        } catch (error2) {}
         this.open = false;
         this.fail('buggy', error);
       } else {
@@ -64,7 +74,9 @@
     onOpenError(event) {
       event.preventDefault();
       this.open = false;
-      const { error } = event.target;
+      const {
+        error
+      } = event.target;
 
       switch (error.name) {
         case 'QuotaExceededError':
@@ -84,14 +96,23 @@
     fail(reason, error) {
       this.cachedDocs = null;
       this.useIndexedDB = false;
-      if (!this.reason) { this.reason = reason; }
-      if (!this.error) { this.error = error; }
-      if (error) { if (typeof console.error === 'function') {
-        console.error('IDB error', error);
-      } }
+      if (!this.reason) {
+        this.reason = reason;
+      }
+      if (!this.error) {
+        this.error = error;
+      }
+      if (error) {
+        if (typeof console.error === 'function') {
+          console.error('IDB error', error);
+        }
+      }
       this.runCallbacks();
       if (error && (reason === 'cant_open')) {
-        Raven.captureMessage(`${error.name}: ${error.message}`, {level: 'warning', fingerprint: [error.name]});
+        Raven.captureMessage(`${error.name}: ${error.message}`, {
+          level: 'warning',
+          fingerprint: [error.name]
+        });
       }
     }
 
@@ -99,7 +120,9 @@
       this.reset();
       this.db();
       app.onQuotaExceeded();
-      Raven.captureMessage('QuotaExceededError', {level: 'warning'});
+      Raven.captureMessage('QuotaExceededError', {
+        level: 'warning'
+      });
     }
 
     onVersionError() {
@@ -107,7 +130,7 @@
       req.onsuccess = event => {
         return this.handleVersionMismatch(event.target.result.version);
       };
-      req.onerror = function(event) {
+      req.onerror = function (event) {
         event.preventDefault();
         return this.fail('cant_open', error);
       };
@@ -123,10 +146,15 @@
     }
 
     buggyIDB(db) {
-      if (this.checkedBuggyIDB) { return; }
+      if (this.checkedBuggyIDB) {
+        return;
+      }
       this.checkedBuggyIDB = true;
       try {
-        this.idbTransaction(db, {stores: $.makeArray(db.objectStoreNames).slice(0, 2), mode: 'readwrite'}).abort(); // https://bugs.webkit.org/show_bug.cgi?id=136937
+        this.idbTransaction(db, {
+          stores: $.makeArray(db.objectStoreNames).slice(0, 2),
+          mode: 'readwrite'
+        }).abort(); // https://bugs.webkit.org/show_bug.cgi?id=136937
         return;
       } catch (error) {
         return error;
@@ -135,39 +163,55 @@
 
     runCallbacks(db) {
       let fn;
-      while ((fn = this.callbacks.shift())) { fn(db); }
+      while ((fn = this.callbacks.shift())) {
+        fn(db);
+      }
     }
 
     onUpgradeNeeded(event) {
       let db;
-      if (!(db = event.target.result)) { return; }
+      if (!(db = event.target.result)) {
+        return;
+      }
 
       const objectStoreNames = $.makeArray(db.objectStoreNames);
 
       if (!$.arrayDelete(objectStoreNames, 'docs')) {
-        try { db.createObjectStore('docs'); } catch (error) {}
+        try {
+          db.createObjectStore('docs');
+        } catch (error) {}
       }
 
       for (let doc of Array.from(app.docs.all())) {
         if (!$.arrayDelete(objectStoreNames, doc.slug)) {
-          try { db.createObjectStore(doc.slug); } catch (error1) {}
+          try {
+            db.createObjectStore(doc.slug);
+          } catch (error1) {}
         }
       }
 
       for (let name of Array.from(objectStoreNames)) {
-        try { db.deleteObjectStore(name); } catch (error2) {}
+        try {
+          db.deleteObjectStore(name);
+        } catch (error2) {}
       }
     }
 
     store(doc, data, onSuccess, onError, _retry) {
-      if (_retry == null) { _retry = true; }
+      if (_retry == null) {
+        _retry = true;
+      }
       this.db(db => {
         if (!db) {
           onError();
           return;
         }
 
-        const txn = this.idbTransaction(db, {stores: ['docs', doc.slug], mode: 'readwrite', ignoreError: false});
+        const txn = this.idbTransaction(db, {
+          stores: ['docs', doc.slug],
+          mode: 'readwrite',
+          ignoreError: false
+        });
         txn.oncomplete = () => {
           if (this.cachedDocs != null) {
             this.cachedDocs[doc.slug] = doc.mtime;
@@ -180,8 +224,7 @@
             this.migrate();
             setTimeout(() => {
               return this.store(doc, data, onSuccess, onError, false);
-            }
-            , 0);
+            }, 0);
           } else {
             onError(event);
           }
@@ -189,7 +232,10 @@
 
         let store = txn.objectStore(doc.slug);
         store.clear();
-        for (let path in data) { const content = data[path]; store.add(content, path); }
+        for (let path in data) {
+          const content = data[path];
+          store.add(content, path);
+        }
 
         store = txn.objectStore('docs');
         store.put(doc.mtime, doc.slug);
@@ -197,28 +243,33 @@
     }
 
     unstore(doc, onSuccess, onError, _retry) {
-      if (_retry == null) { _retry = true; }
+      if (_retry == null) {
+        _retry = true;
+      }
       this.db(db => {
         if (!db) {
           onError();
           return;
         }
 
-        const txn = this.idbTransaction(db, {stores: ['docs', doc.slug], mode: 'readwrite', ignoreError: false});
+        const txn = this.idbTransaction(db, {
+          stores: ['docs', doc.slug],
+          mode: 'readwrite',
+          ignoreError: false
+        });
         txn.oncomplete = () => {
           if (this.cachedDocs != null) {
             delete this.cachedDocs[doc.slug];
           }
           onSuccess();
         };
-        txn.onerror = function(event) {
+        txn.onerror = function (event) {
           event.preventDefault();
           if (((txn.error != null ? txn.error.name : undefined) === 'NotFoundError') && _retry) {
             this.migrate();
             setTimeout(() => {
               return this.unstore(doc, onSuccess, onError, false);
-            }
-            , 0);
+            }, 0);
           } else {
             onError(event);
           }
@@ -245,14 +296,17 @@
           return;
         }
 
-        const txn = this.idbTransaction(db, {stores: ['docs'], mode: 'readonly'});
+        const txn = this.idbTransaction(db, {
+          stores: ['docs'],
+          mode: 'readonly'
+        });
         const store = txn.objectStore('docs');
 
         const req = store.get(doc.slug);
-        req.onsuccess = function() {
+        req.onsuccess = function () {
           fn(req.result);
         };
-        req.onerror = function(event) {
+        req.onerror = function (event) {
           event.preventDefault();
           fn(false);
         };
@@ -260,7 +314,9 @@
     }
 
     cachedVersion(doc) {
-      if (!this.cachedDocs) { return; }
+      if (!this.cachedDocs) {
+        return;
+      }
       return this.cachedDocs[doc.slug] || false;
     }
 
@@ -277,19 +333,22 @@
           return;
         }
 
-        const txn = this.idbTransaction(db, {stores: ['docs'], mode: 'readonly'});
-        txn.oncomplete = function() {
+        const txn = this.idbTransaction(db, {
+          stores: ['docs'],
+          mode: 'readonly'
+        });
+        txn.oncomplete = function () {
           fn(result);
         };
         const store = txn.objectStore('docs');
         var result = {};
 
-        docs.forEach(function(doc) {
+        docs.forEach(function (doc) {
           const req = store.get(doc.slug);
-          req.onsuccess = function() {
+          req.onsuccess = function () {
             result[doc.slug] = req.result;
           };
-          req.onerror = function(event) {
+          req.onerror = function (event) {
             event.preventDefault();
             result[doc.slug] = false;
           };
@@ -298,9 +357,13 @@
     }
 
     cachedVersions(docs) {
-      if (!this.cachedDocs) { return; }
+      if (!this.cachedDocs) {
+        return;
+      }
       const result = {};
-      for (let doc of Array.from(docs)) { result[doc.slug] = this.cachedVersion(doc); }
+      for (let doc of Array.from(docs)) {
+        result[doc.slug] = this.cachedVersion(doc);
+      }
       return result;
     }
 
@@ -335,14 +398,21 @@
           return;
         }
 
-        const txn = this.idbTransaction(db, {stores: [entry.doc.slug], mode: 'readonly'});
+        const txn = this.idbTransaction(db, {
+          stores: [entry.doc.slug],
+          mode: 'readonly'
+        });
         const store = txn.objectStore(entry.doc.slug);
 
         const req = store.get(entry.dbPath());
-        req.onsuccess = function() {
-          if (req.result) { onSuccess(req.result); } else { onError(); }
+        req.onsuccess = function () {
+          if (req.result) {
+            onSuccess(req.result);
+          } else {
+            onError();
+          }
         };
-        req.onerror = function(event) {
+        req.onerror = function (event) {
           event.preventDefault();
           onError();
         };
@@ -351,10 +421,15 @@
     }
 
     loadDocsCache(db) {
-      if (this.cachedDocs) { return; }
+      if (this.cachedDocs) {
+        return;
+      }
       this.cachedDocs = {};
 
-      const txn = this.idbTransaction(db, {stores: ['docs'], mode: 'readonly'});
+      const txn = this.idbTransaction(db, {
+        stores: ['docs'],
+        mode: 'readonly'
+      });
       txn.oncomplete = () => {
         setTimeout(this.checkForCorruptedDocs, 50);
       };
@@ -362,11 +437,13 @@
       const req = txn.objectStore('docs').openCursor();
       req.onsuccess = event => {
         let cursor;
-        if (!(cursor = event.target.result)) { return; }
+        if (!(cursor = event.target.result)) {
+          return;
+        }
         this.cachedDocs[cursor.key] = cursor.value;
         cursor.continue();
       };
-      req.onerror = function(event) {
+      req.onerror = function (event) {
         event.preventDefault();
       };
     }
@@ -384,7 +461,9 @@
           }
           return result;
         })());
-        if (docs.length === 0) { return; }
+        if (docs.length === 0) {
+          return;
+        }
 
         for (var slug of Array.from(docs)) {
           if (!app.docs.findBy('slug', slug)) {
@@ -401,14 +480,22 @@
           return;
         }
 
-        const txn = this.idbTransaction(db, {stores: docs, mode: 'readonly', ignoreError: false});
+        const txn = this.idbTransaction(db, {
+          stores: docs,
+          mode: 'readonly',
+          ignoreError: false
+        });
         txn.oncomplete = () => {
-          if (this.corruptedDocs.length > 0) { setTimeout(this.deleteCorruptedDocs, 0); }
+          if (this.corruptedDocs.length > 0) {
+            setTimeout(this.deleteCorruptedDocs, 0);
+          }
         };
 
         for (let doc of Array.from(docs)) {
           txn.objectStore(doc).get('index').onsuccess = event => {
-            if (!event.target.result) { this.corruptedDocs.push(event.target.source.name); }
+            if (!event.target.result) {
+              this.corruptedDocs.push(event.target.source.name);
+            }
           };
         }
       });
@@ -417,14 +504,23 @@
     deleteCorruptedDocs() {
       this.db(db => {
         let doc;
-        const txn = this.idbTransaction(db, {stores: ['docs'], mode: 'readwrite', ignoreError: false});
+        const txn = this.idbTransaction(db, {
+          stores: ['docs'],
+          mode: 'readwrite',
+          ignoreError: false
+        });
         const store = txn.objectStore('docs');
         while ((doc = this.corruptedDocs.pop())) {
           this.cachedDocs[doc] = false;
           store.delete(doc);
         }
       });
-      Raven.captureMessage('corruptedDocs', {level: 'info', extra: { docs: this.corruptedDocs.join(',') }});
+      Raven.captureMessage('corruptedDocs', {
+        level: 'info',
+        extra: {
+          docs: this.corruptedDocs.join(',')
+        }
+      });
     }
 
     shouldLoadWithIDB(entry) {
@@ -435,12 +531,12 @@
       app.lastIDBTransaction = [options.stores, options.mode];
       const txn = db.transaction(options.stores, options.mode);
       if (options.ignoreError !== false) {
-        txn.onerror = function(event) {
+        txn.onerror = function (event) {
           event.preventDefault();
         };
       }
       if (options.ignoreAbort !== false) {
-        txn.onabort = function(event) {
+        txn.onabort = function (event) {
           event.preventDefault();
         };
       }
@@ -448,9 +544,11 @@
     }
 
     reset() {
-      try { if (typeof indexedDB !== 'undefined' && indexedDB !== null) {
-        indexedDB.deleteDatabase(NAME);
-      } } catch (error) {}
+      try {
+        if (typeof indexedDB !== 'undefined' && indexedDB !== null) {
+          indexedDB.deleteDatabase(NAME);
+        }
+      } catch (error) {}
     }
 
     useIndexedDB() {
